+++ CHANGELOG.md
@@ -3,10 +3,14 @@
 
 ## Added
 
+### v0.2.2
+- Fixed the pause/resume icon problem
+- Fixed setup function. Can now change values
 
-### v0.2.0  
+### v0.2.1
 
 - Extension Works
+- Problems in pause/resume icon
 
 
 #### v0.01 - v0.1.9